[Amount of money missing][green] It should calculate how much money is missing exactly.
- Replace float numbers by BigDecimals for calculation.

+++ src/main/java/software/works/machine/CustomerOrder.java
@@ -1,21 +1,23 @@
 package software.works.machine;
 
+import java.math.BigDecimal;
+
 public class CustomerOrder {
 
     private Drink drink;
     private int numberOfSugars;
-    private float amountOfMoney;
+    private BigDecimal amountOfMoney;
 
     public CustomerOrder(Drink drink) {
         this.drink = drink;
     }
 
-    public CustomerOrder(Drink drink, float amountOfMoney) {
+    public CustomerOrder(Drink drink, BigDecimal amountOfMoney) {
         this.drink = drink;
         this.amountOfMoney = amountOfMoney;
     }
 
-    public CustomerOrder(Drink drink, int numberOfSugars, float amountOfMoney) {
+    public CustomerOrder(Drink drink, int numberOfSugars, BigDecimal amountOfMoney) {
         this.drink = drink;
         this.numberOfSugars = numberOfSugars;
         this.amountOfMoney = amountOfMoney;
@@ -29,7 +31,7 @@ public class CustomerOrder {
         return numberOfSugars;
     }
 
-    public float getAmountOfMoney() {
+    public BigDecimal getAmountOfMoney() {
         return amountOfMoney;
     }
